deps/v8/test/webkit/fast/js/Promise-then-without-callbacks-expected.txt

Promise.prototype.then should work without callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
